hyperv: enable PCIe pass-through (a.k.a. Discrete Device Assignment)
The feature enables us to pass through physical PCIe devices to FreeBSD VM
running on Hyper-V (Windows Server 2016) to get near-native performance with
low CPU utilization.

The patch implements a PCI bridge driver to support the feature:

1) the pcib driver talks to the host to discover device(s) and presents
the device(s) to FreeBSD's pci driver via PCI configuration space (note:
to access the configuration space, we don't use the standard I/O port
0xCF8/CFC method; instead, we use an MMIO-based method supplied by Hyper-V,
which is very similar to the 0xCF8/CFC method).

2) the pcib driver allocates resources for the device(s) and initialize
the related BARs, when the device driver's attach method is invoked;

3) the pcib driver talks to the host to create MSI/MSI-X interrupt
remapping between the guest and the host;

4) the pcib driver supports device hot add/remove.